Extract functions and types from Switch capability
Related to #44
