--- src/hubitat-capabilities/bulb.capability.ts
@@ -1,9 +1,0 @@
-/**
- * @packageDocumentation
- * @module HubitatCapabilities
- */
-
-export { isAnySwitchOn as isAnyBulbOn } from './switch.capability';
-export { isSwitchOn as isBulbOn } from './switch.capability';
-export { setSwitch as setBulb } from './switch.capability';
-export { toggleSwitch as toggleBulb } from './switch.capability';
