+++ src/hubitat-capabilities/index.ts
@@ -12,6 +12,8 @@ export * from './functions/get-thermostat-operating-state.function';
 export * from './functions/get-thermostat-schedule.function';
 export * from './functions/get-thermostat-setpoint.function';
 
+export * from './functions/is-any-on.function';
+export * from './functions/is-on.function';
 export * from './functions/is-open.function';
 
 export * from './functions/set-alarm-setting.function';
@@ -19,6 +21,7 @@ export * from './functions/set-cooling-setpoint.function';
 export * from './functions/set-heating-setpoint.function';
 export * from './functions/set-supported-thermostat-fan-modes.function';
 export * from './functions/set-supported-thermostat-modes.function';
+export * from './functions/set-switch.function';
 export * from './functions/set-thermostat-fan-mode.function';
 export * from './functions/set-thermostat-mode.function';
 export * from './functions/set-thermostat-operating-state.function';
@@ -28,11 +31,14 @@ export * from './functions/set-thermostat-setpoint.function';
 export * from './functions/start-level-change.function';
 export * from './functions/stop-level-change.function';
 
+export * from './functions/toggle.function';
+
 /* TYPES */
 
 export * from './types/alarm-setting.type';
 export * from './types/level-change-direction.type';
 export * from './types/open-status.type';
+export * from './types/switch-status.type';
 export * from './types/thermostat-fan-modes.type';
 export * from './types/thermostat-modes.type';
 export * from './types/thermostat-operating-modes.type';
