--- src/hubitat-capabilities/light.capability.ts
@@ -1,9 +1,0 @@
-/**
- * @packageDocumentation
- * @module HubitatCapabilities
- */
-
-export { isAnySwitchOn as isAnyLightOn } from './switch.capability';
-export { isSwitchOn as isLightOn } from './switch.capability';
-export { setSwitch as setLight } from './switch.capability';
-export { toggleSwitch as toggleLight } from './switch.capability';
