--- src/hubitat-capabilities/outlet.capability.ts
@@ -1,9 +1,0 @@
-/**
- * @packageDocumentation
- * @module HubitatCapabilities
- */
-
-export { isAnySwitchOn as isAnyOutletOn } from './switch.capability';
-export { isSwitchOn as isOutletOn } from './switch.capability';
-export { setSwitch as setOutlet } from './switch.capability';
-export { toggleSwitch as toggleOutlet } from './switch.capability';
