--- src/hubitat-capabilities/relay-switch.capability.ts
@@ -1,9 +1,0 @@
-/**
- * @packageDocumentation
- * @module HubitatCapabilities
- */
-
-export { isAnySwitchOn as isAnyRelaySwitchOn } from './switch.capability';
-export { isSwitchOn as isRelaySwitchOn } from './switch.capability';
-export { setSwitch as setRelaySwitch } from './switch.capability';
-export { toggleSwitch as toggleRelaySwitch } from './switch.capability';
