+++ src/hubitat-capabilities/types/switch-status.type.ts
@@ -1,0 +1,9 @@
+/**
+ * @packageDocumentation
+ * @module HubitatCapabilities
+ */
+
+/**
+ * Switch status.
+ */
+export type SwitchStatus = 'off' | 'on';
